Move cpp_version_in_use() from cast.h to pybind11_tests.cpp
