+++ tests/pybind11_tests.cpp
@@ -62,10 +62,23 @@ void bind_ConstructorStats(py::module_ &m) {
         });
 }
 
+const char *cpp_version_in_use() {
+    return
+#if defined(PYBIND11_CPP20)
+        "C++20";
+#elif defined(PYBIND11_CPP17)
+        "C++17";
+#elif defined(PYBIND11_CPP14)
+        "C++14";
+#else
+        "C++11";
+#endif
+}
+
 PYBIND11_MODULE(pybind11_tests, m) {
     m.doc() = "pybind11 test module";
 
-    m.attr("cpp_version_in_use") = py::detail::cpp_version_in_use();
+    m.attr("cpp_version_in_use") = cpp_version_in_use();
     m.attr("PYBIND11_INTERNALS_ID") = PYBIND11_INTERNALS_ID;
 
     bind_ConstructorStats(m);
